Removed AWS key related functions.
Renamed file to env.
